Added DietGoalController, with getAll method, and modified calculateDietPlan to search for the dietgoals in database

+++ SmartMealFinderB/src/main/java/hu/project/smartmealfinderb/Request/DietPlanRequest.java
@@ -6,8 +6,8 @@ import lombok.Data;
 public class DietPlanRequest {
 
     private String sex;
-    private float weight;
-    private float height;
+    private double weight;
+    private double height;
     private int age;
     private int activityLevel;
     private int goalType;
